refactor disease content
