+++ static/sass/object/component/_disease-side-nav.scss
@@ -38,7 +38,7 @@
         background-color: rgba(22, 35, 78, 0.2);
         color: var(--color-secondary);
       }
-      .brc,
+      .bio-resource,
       .variant {
         font-size: 14px;
         cursor: pointer;
